Vacation_Website/requirements.txt
django==5.2.3
psycopg2-binary==2.9.10
psycopg==3.2.9
djangorestframework==3.16.0
django-allauth==65.9.0
requests==2.32.4
pillow==11.2.1
python-decouple==3.8
django-cors-headers==4.7.0
